Simple web page which does addition of 2 numbers.
Uses a spark static files feature to read static files from
src/main/resources -- useful for html, css and javascript files.

The spark handler is a simple post method which reads two parameters (a
and b), sums them together and returns the sum as a result.

The static page (accessible from http://localhost:4567) presents a form
with two inputs and a button. The inputs are sent (via ajax) to the
spark handler and the result (or error) is added to page.